Commenting code and attempted Reciprocal Collision Avoidance
I attempted to create a Reciprocal Collision Avoidance system for the units so they would not hit each other. However, I failed to get the system fully working (this can be seen in the unit script). So the only method for the units not clipping was to use their colliders, which in itself had its own problem of the units floating away when they would reach the target and collided.